Migrate generate QR functionality
